+++ src/fe_vm.cpp
@@ -1665,8 +1665,8 @@ bool FeVM::on_new_layout()
 
 	fe.Func( _SC("add_clone"), &FeVM::cb_add_clone_presentable );
 
-	fe.Overload<FeText* (*)(const char *, int, int, int, int)>(_SC("add_text"), &FeVM::cb_add_text);
-	fe.Func<FeListBox* (*)(int, int, int, int)>(_SC("add_listbox"), &FeVM::cb_add_listbox);
+	fe.Overload<FeText* (*)(const char *, float, float, float, float)>(_SC("add_text"), &FeVM::cb_add_text);
+	fe.Func<FeListBox* (*)(float, float, float, float)>(_SC("add_listbox"), &FeVM::cb_add_listbox);
 	fe.Func<FeRectangle* (*)(float, float, float, float)>(_SC("add_rectangle"), &FeVM::cb_add_rectangle);
 	fe.Overload<FeImage* (*)(float, float, float, float, int, int)>(_SC("add_surface"), &FeVM::cb_add_surface);
 	fe.Overload<FeImage* (*)(float, float, float, float)>(_SC("add_surface"), &FeVM::cb_add_surface);
@@ -2935,7 +2935,7 @@ Sqrat::Object FeVM::cb_add_clone_presentable( FeBasePresentable *o )
 	return ret_obj;
 }
 
-FeText* FeVM::cb_add_text(const char *n, int x, int y, int w, int h )
+FeText* FeVM::cb_add_text(const char *n, float x, float y, float w, float h )
 {
 	HSQUIRRELVM vm = Sqrat::DefaultVM::Get();
 	FeVM *fev = (FeVM *)sq_getforeignptr( vm );
@@ -2951,7 +2951,7 @@ FeText* FeVM::cb_add_text(const char *n, int x, int y, int w, int h )
 	return ret;
 }
 
-FeListBox* FeVM::cb_add_listbox(int x, int y, int w, int h )
+FeListBox* FeVM::cb_add_listbox(float x, float y, float w, float h )
 {
 	HSQUIRRELVM vm = Sqrat::DefaultVM::Get();
 	FeVM *fev = (FeVM *)sq_getforeignptr( vm );
